#11 디렉터리 구조 개선, 그리드 전체 html 에 적용, 카테고리 메뉴 생성
## 디렉터리 구조 개선
- 'src' - 전체 소스 코드를 'src' 디렉터리에서 관리.
- 프로젝트 설정 파일 등은 'src' 밖에 놓음.
- 'src/clint' - 브라우저 측의 최상위 컴포넌트 (추후 서버사이드 렌더링 구현을 위해 따로 분리)
- 'src/component' - 컴포넌트 파일 위치.
- 'src/pages' - 각 라우터들 (페이지 별).
- 'src/shared' - 서버와 클라이언트에서 공용으로 사용되는 컴포넌트.
- 'src/static' - 정적 파일(stylesheet, image, js... / src 바깥으로 빼야할듯?).

## 그리드 전체 html 에 적용
- 'src/client/root.jsx' - 최상위 컴포넌트에 전체 그리드 적용.

## 카테고리 메뉴 생성
- 'src/component/Menu.jsx' - 각 페이지 이동 할 수 있는 컴포넌트 생성.
