--- client/categories/sample.jsx
@@ -1,16 +1,0 @@
-import React , { Component } from 'react';
-import {Button} from "antd";
-
-class Categorie extends Component {
-
-  render() {
-    return (
-        <>
-          <h1>Categories</h1>
-          <Button type="primary">Example button</Button>
-        </>
-    );
-  }
-}
-
-export default Categorie;
