+++ client/index.html
@@ -6,6 +6,6 @@
 </head>
 <body>
 <div id="root"></div>
-<script src="./app.jsx"></script>
+<script src="../app.jsx"></script>
 </body>
 </html>
