--- client/info/sample.jsx
@@ -1,1 +1,0 @@
-aasd
