+++ client/package.json
@@ -25,6 +25,7 @@
     "@babel/preset-env": "^7.5.4",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.6",
+    "babel-plugin-import": "^1.12.0",
     "react-hot-loader": "^4.12.6",
     "webpack": "^4.35.3",
     "webpack-cli": "^3.3.6",
