+++ client/src/component/Menu.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+// hooks
+const Menu = () => {
+  return (
+    <div>
+      <ul>
+        <li><Link to="/">Home</Link></li>
+        <li><Link to="/report">report</Link></li>
+      </ul>
+    </div>
+  );
+}
+
+export default Menu;
