+++ client/src/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import Root from './client/root';
+
+ReactDOM.render(<Root />, document.getElementById('root'));
