+++ client/src/pages/index.jsx
@@ -1,0 +1,2 @@
+export { default as Home } from './home';
+export { default as Report } from './report';
