+++ client/src/static/style/report.css
@@ -1,0 +1,3 @@
+.container {
+  background-color: blue;
+}
